requirements.txt
python-dotenv
streamlit
streamlit-cookies-controller
langchain>=0.0.310
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-openai>=0.0.3
qdrant-client
openai
pdfkit
python-docx
pyyaml
streamlit-authenticator
reportlab
requests
urllib3
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.5.0
gitdb==4.0.12
gitpython==3.1.44
greenlet==3.1.1
grpcio==1.68.1
grpcio-tools==1.68.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.0.1
idna==3.10
jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.23.3
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.20.1
numpy==2.2.1
orjson==3.10.13
packaging==24.2
pandas==2.2.3
pillow==11.1.0
portalocker==2.10.1
propcache==0.2.1
pyarrow==18.1.0
pydantic>=2.4.0
pydantic-core>=2.10.0
pydantic-settings>=2.0.0
pydeck==0.9.1
pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.11.6
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.36
tenacity==9.0.0
tiktoken>=0.5.2
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
watchdog==6.0.0
yarl==1.18.0
redis==5.0.0
locust==2.24.1
langchain_qdrant==0.2.0
psycopg2-binary==2.9.10
pymysql
DBUtils==3.0.3
protobuf<6.0.0,>=4.25.0
gunicorn==21.2.0